easysetup: extract methods from http tcp server
extract socket related methods from http tcp server
