+++ src/easysetup/CMakeLists.txt
@@ -17,7 +17,8 @@ else()
     set(EASYSETUP_D2D_SOURCES
         "${EASYSETUP_D2D_SOURCES}"
         http/tcp/iot_easysetup_d2d_tcp.c
-        http/tcp/iot_easysetup_http_tcp.c
+        http/tcp/iot_easysetup_httpd.c
+        http/tcp/iot_easysetup_http_socket.c
     )
 endif()
